Call and Put pricing Plotted

+++ esercizio_3/rng/GeometricBrownianMotion.cpp
@@ -6,25 +6,25 @@
 #include "random.h"
 
 
-float GBM(float S, float T, float mu, float sigma , Random & rnd){
+double GBM(double S, double T, double mu, double sigma , Random & rnd){
 	return S*exp( (mu-0.5 * sigma*sigma)*T + sigma*rnd.Gauss(0,1)*sqrt(T));
 }
 
 // Black-Scholes:
 
-float N(float x){
+double N(double x){
 	return 0.5 * (1 + erf( x/sqrt(2) ) );
 }
 
-float call_option(float S, float T, float K, float sigma, float r){
+double call_option(double S, double T, double K, double sigma, double r){
 
-	float d1 = ( log(S/K) + (r + 0.5*sigma*sigma*T) ) / (sigma*sqrt(T));
-	float d2 = d1 - sigma*sqrt(T);
+	double d1 = ( log(S/K) + (r + 0.5*sigma*sigma*T) ) / (sigma*sqrt(T));
+	double d2 = d1 - sigma*sqrt(T);
 	return S*N(d1) - K*exp(-r*T)*N(d2);
 }
 
-float put_option(float S, float T, float K, float sigma, float r){
-	float d1 = ( log(S/K) + (r + 0.5*sigma*sigma*T) ) / (sigma*sqrt(T));
-	float d2 = d1 - sigma*sqrt(T);
+double put_option(double S, double T, double K, double sigma, double r){
+	double d1 = ( log(S/K) + (r + 0.5*sigma*sigma*T) ) / (sigma*sqrt(T));
+	double d2 = d1 - sigma*sqrt(T);
 	return S*(N(d1)-1) - K*exp(-r*T)*(N(d2)-1);
 }
